Refactor button sizes and themes
